Remade Makefile to support file cacheing
This makefile will now only recompile files that need to be recompiled.

There are a lot of issues that need to be solved before this can be merged with main

I need to include -DLWIP_PROVIDE_ERRNO=22 as a preprocessor flag or else some file can't compile

The Dynamic finding of ARDUINO_PATH is broken and hardcoded to my computer. Sorry!!

I had to take the TeensyDebug library out of the library folder (its temporarily stored in the root of the repo) because this would also cause compile errors

in the main, the compiler can't find git.info.h

other things are commented out in main because of this and also because of teensyDebug not working